// front/src/api/models/DiscreteMeta.ts
/* tslint:disable */
/* eslint-disable */
/**
 * demo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DiscreteMeta
 */
export interface DiscreteMeta {
    /**
     * 
     * @type {string}
     * @memberof DiscreteMeta
     */
    name: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof DiscreteMeta
     */
    values: Array<string>;
}

export function DiscreteMetaFromJSON(json: any): DiscreteMeta {
    return DiscreteMetaFromJSONTyped(json, false);
}

export function DiscreteMetaFromJSONTyped(json: any, ignoreDiscriminator: boolean): DiscreteMeta {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'values': json['values'],
    };
}

export function DiscreteMetaToJSON(value?: DiscreteMeta | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'values': value.values,
    };
}